CT - Add rudimentary caching for basic protection when live

+++ lib/api_interfaces.py
@@ -1,5 +1,6 @@
 import requests
 import datetime
+import time
 from pathlib import Path
 import json
 from Scorebook import Scorebook
@@ -11,6 +12,7 @@ DATA_DIR = Path('data/')
 GAME_DIR = DATA_DIR/'game'
 DAY_DIR = DATA_DIR/'day'
 ALL_PATHS = [DATA_DIR, GAME_DIR, DAY_DIR]
+RATE_LIMIT = 5 #SECONDS
 
 def init_dir():
     for p in ALL_PATHS:
@@ -20,18 +22,24 @@ def init_dir():
 
 
 def cached_download_json(cache_path, url, force_download = False):
+# TODO - Weird logical structure
     if not cache_path.exists() or force_download:
-        print(f"No cache hit for {cache_path}. Downloading")
-        try:
-            resp = requests.get(url)
-        except Exception as e:
-            print(f"Failed downloading ids for date {cache_path}")
-            print(f"Exception - {str(e)}")
-            raise
+        if cache_path.exists() and (time.time() - cache_path.stat().st_mtime) < RATE_LIMIT:
+            print(f"Rate limited for {cache_path}. Using cache")
+            with cache_path.open() as f:
+                djson = json.load(f)
+        else:
+            print(f"No cache hit for {cache_path}. Downloading")
+            try:
+                resp = requests.get(url)
+            except Exception as e:
+                print(f"Failed downloading ids for date {cache_path}")
+                print(f"Exception - {str(e)}")
+                raise
 
-        djson = resp.json()
-        with cache_path.open('w') as f:
-            json.dump(djson, f)
+            djson = resp.json()
+            with cache_path.open('w') as f:
+                json.dump(djson, f)
     else:
         print(f"Using cache hit for {cache_path}")
         with cache_path.open() as f:
@@ -82,5 +90,5 @@ def get_game(gameid, refresh = False):
     away.save(str(away_file))
     #except:
         #return False
-
-    return {"home": home_file, "away": away_file}
+    age = time.time() - cache_f.stat().st_mtime
+    return {"home": home_file, "away": away_file, "age": age}
